docker-compose-demo/api/requirements.txt
Flask==2.0.1
flask-cors==3.0.10
redis==4.0.2
gunicorn==20.1.0
psycopg2-binary==2.9.9
Werkzeug==2.0.1 